# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Village Swimming Pool Reservation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List reservations
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationListResponse"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: Conflict or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get reservation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Update reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    delete:
      operationId: deleteReservation
      tags: [reservations]
      summary: Cancel reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /reservations/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [reservations]
      summary: Get available time slots for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
  /reservations/my:
    get:
      operationId: getMyReservations
      tags: [reservations]
      summary: Get current user's reservations
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: My reservations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationListResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        houseNumber:
          type: string
        phone:
          type: string
        email:
          type: string
        username:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - houseNumber
        - phone
        - email
        - username
        - role
        - createdAt
...
    ReservationInput:
      type: object
      properties:
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        numberOfPeople:
          type: integer
        notes:
          type: ["string", "null"]
      required:
        - date
        - startTime
        - endTime
        - numberOfPeople
    ReservationUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
        numberOfPeople:
          type: integer
    Reservation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        numberOfPeople:
          type: integer
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - userId
        - date
        - startTime
        - endTime
        - numberOfPeople
        - status
        - createdAt
        - user
    ReservationListResponse:
      type: object
      properties:
        reservations:
          type: array
          items:
            $ref: "#/components/schemas/Reservation"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - reservations
        - total
        - page
        - totalPages
    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        available:
          type: boolean
        currentPeople:
          type: integer
        maxPeople:
          type: integer
      required:
        - startTime
        - endTime
        - available
        - currentPeople
        - maxPeople
...